refactor(notebook): move mutation gate to shared surface
